sexcs/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sec-quest"
version = "1.0.0"
description = "PR Code Review RL environment for OpenEnv — find planted bugs in code diffs"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "websockets>=11.0",
    "requests>=2.25.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "httpx>=0.24.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["sec_quest*", "server*"]